通过不同group可以获取同一个Topic的消息，实现点赞、取消点赞日增量数据同步入库（临时表）

+++ linker-data-align/src/main/java/com/aus/linker/data/align/constant/TableConstants.java
@@ -7,7 +7,7 @@ public class TableConstants {
      */
     private static final String TABLE_NAME_SEPARATOR = "_";
 
-    public static String buildTableNameSuffix(String date, int hashKey) {
+    public static String buildTableNameSuffix(String date, long hashKey) {
         // 拼接完整的表名
         return date + TABLE_NAME_SEPARATOR + hashKey;
     }
